app/requirements.txt
aniso8601==9.0.1
beautifulsoup4==4.10.0
cachetools==4.2.4
certifi==2021.10.8
charset-normalizer==2.0.9       
click==8.0.3
colorama==0.4.4
Flask==2.0.2
Flask-RESTful==0.3.9
google==3.0.0
google-api-core==2.3.0
google-auth==2.3.3
google-cloud-core==2.2.1        
google-cloud-storage==1.43.0    
google-crc32c==1.3.0
google-resumable-media==2.1.0   
googleapis-common-protos==1.54.0
idna==3.3
itsdangerous==2.0.1
Jinja2==3.0.3
MarkupSafe==2.0.1
protobuf==3.19.1
psycopg2==2.9.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
python-dotenv==0.19.2
pytz==2021.3
requests==2.26.0
rsa==4.8
six==1.16.0
soupsieve==2.3.1
urllib3==1.26.7
Werkzeug==2.0.2
elasticsearch==7.15.2